Included final challenge solution
Included both the IPython Notebook and the Python file of the final solution to the challenge. This includes a tested and working solution to figuring out how to figure out the best clients to sell to in order to maximize profits while using minimal computing power. However, I did have trouble importing the given dataset as a .csv or as an excel file so I set up the dataset manually in Python. If I were to work on the solution more, I would try to figure out a more efficient method of importing the data rather than typing it all out by hand.